// src/pages/dashboard/types.ts
export interface ITarget {
  refId: string;
  expr: string; // promQL
  legendFormat: string;
  time?: IRawTimeRange; // 固定时间范围
  step?: number; // 固定时间间隔
  query?: {
    index: string;
    filters: string;
    values: {
      func: string;
    }[];
  };
  legend?: string;
  instant?: boolean;
}